requirements.txt
asgiref==3.2.10
boto3==1.15.7
botocore==1.18.7
Brotli==1.0.7
colorama==0.4.3
Django==3.0.8
django-material==1.7.0
django-materializecss-form==1.1.17
django-storages==1.10.1
django-tinymce4-lite==1.8.0
gunicorn==20.0.4
jmespath==0.10.0
jsmin==2.2.2
Pillow==7.2.0
pypng==0.0.20
PyQRCode==1.2.1
python-dateutil==2.8.1
pytz==2020.1
qrcode==6.1
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
whitenoise==5.1.0
